new file:   checkout-process.php 	modified:   config/connect.php 	modified:   config/filter.php 	modified:   page/admin/index.php 	modified:   page/checkout.php 	modified:   page/checkout_confirm.php 	modified:   page/home.php 	modified:   page/list.php 	modified:   page/product/add.php 	new file:   page/product/data/display.php 	new file:   page/product/data/edit.php 	deleted:    page/product/data/list.php 	modified:   page/product/edit.php 	modified:   page/product/list.php 	modified:   page/product/view.php 	modified:   page/profile/edit.php 	modified:   page/profile/view.php 	modified:   page/register.php 	new file:   page/report/review/list.php 	modified:   page/review/add.php 	modified:   page/wishlist/list.php

+++ page/checkout_confirm.php
@@ -3,8 +3,34 @@ $result = get('user', 'WHERE email="' . $email . '"');
 $data = mysqli_fetch_assoc($result);
 $user_id = $data['user_id'];
 
-$query = "DELETE FROM cart WHERE user_id=".$user_id;
+// SOLD
+$cart_get = get('cart', 'WHERE user_id=' . $user_id . ' GROUP BY product_id', '*,SUM(quantity) AS total_quantity');
 
-if (mysqli_query($connect, $query)) {
-    echo "<script>window.location.href = 'index.php?page=cart_list'</script>";
+foreach ($cart_get as $data) {
+    $product_id = $data['product_id'];
+    $total_quantity = $data['total_quantity'];
+
+    $product_get = get('product', 'WHERE product_id=' . $product_id);
+    $product_data = mysqli_fetch_assoc($product_get);
+
+    $sold_old = $product_data['sold'];
+    $sold_new = $sold_old + $total_quantity;
+
+    $query = 'UPDATE product SET sold="' . $sold_new . '" WHERE product_id=' . $product_id;
+    mysqli_query($connect, $query);
+}
+
+// HISTORY
+$result = insert('history', [
+    'user_id' => $user_id,
+    'product_id' => $product_id,
+    'date' => date("d-m-Y")
+]);
+
+if ($result) {
+    $query = "DELETE FROM cart WHERE user_id=" . $user_id;
+
+    if (mysqli_query($connect, $query)) {
+        echo "<script>window.location.href = 'index.php?page=cart_list'</script>";
+    }
 }
